Feat: added a StoreIfAbsent method
Added a StoreIfAbsent method that stores an object if and only if it did not exist in the database (or the cache) before. In contrast to the "ComputeIfAbsent" method, this method does not retrieve the currently stored object and therefore does not access the value log. It is therefore offers a very fast way to only store things once with the ability to retrieve if the object was added the first time.

At the same time I fixed a bug, where an entry would not have been updated by "ComputeIfAbsent" if a previous Load retrieved a missing entry.